Rewrite http examples as jest spec (#370)
Rewrite of the http client sample code as jest spec.

This is more elegant for testing the samples in CI as the test server automatically starts and shuts down.

Also, the spec files should be a bit more self documenting as they have proper names. The user can copy/paste example code from the spec file.